Add hook to fetch reports
